More reorg into models
