+++ ramlfications/models/__init__.py
@@ -1,19 +1,18 @@
 # -*- coding: utf-8 -*-
 # Copyright (c) 2016 Spotify AB
 
-from .raml import RootNodeDataType
+from .raml import BaseRootNode, DataTypeNode, RAML_VERSION_LOOKUP
 from .resource_types import ResourceTypeNode
 from .resources import ResourceNode
-from .root import BaseRootNode, RAML_ROOT_LOOKUP
 from .security import SecuritySchemeNode
 from .traits import TraitNode
 
 __all__ = [
     "BaseRootNode",
-    "RAML_ROOT_LOOKUP",
+    "DataTypeNode",
+    "RAML_VERSION_LOOKUP",
     "ResourceTypeNode",
     "ResourceNode",
-    "RootNodeDataType",
     "SecuritySchemeNode",
     "TraitNode",
 ]
