+++ ramlfications/parser/types.py
@@ -1,11 +1,11 @@
 # -*- coding: utf-8 -*-
 # Copyright (c) 2015 The Ramlfications developers
 
-from ramlfications.models import RootNodeDataType
-from ramlfications.types import create_type
+from ramlfications.models import DataTypeNode
+from ramlfications.models.data_types import create_type
 
 
-def create_root_data_type(raml):
+def create_root_data_type(raml, name=None):
     """
     Creates a :py:class:`.raml.RootNodeDataType` based off of the RAML's root\
         section.
@@ -15,8 +15,8 @@ def create_root_data_type(raml):
         attributes set
     """
 
-    return RootNodeDataType(
+    return DataTypeNode(
         raw=raml,
         raml_version=raml._raml_version,
-        type=create_type(None, raml)
+        type=create_type(name, raml)
     )
