--- ramlfications/raml.py
@@ -1,4 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright (c) 2015 Spotify AB
-
-from __future__ import absolute_import, division, print_function
