+++ tests/v020tests/unit/parser/test_parser.py
@@ -11,7 +11,7 @@ from ramlfications import parser as pw
 from ramlfications.parser.parser import RootParser
 from ramlfications.config import setup_config
 from ramlfications.models import TraitNode, ResourceTypeNode
-from ramlfications.models.root import RootNodeAPI08
+from ramlfications.models.raml import RAML08
 from ramlfications.utils import load_file
 
 from tests.base import EXAMPLES
@@ -35,11 +35,11 @@ def root():
 def test_parse_raml(loaded_raml):
     config = setup_config(EXAMPLES + "test-config.ini")
     root = pw.parse_raml(loaded_raml, config)
-    assert isinstance(root, RootNodeAPI08)
+    assert isinstance(root, RAML08)
 
 
 def test_create_root(root):
-    assert isinstance(root, RootNodeAPI08)
+    assert isinstance(root, RAML08)
 
 
 def test_base_uri(root):
