+++ tests/v020tests/unit/test_init.py
@@ -6,7 +6,7 @@ import os
 import pytest
 
 from ramlfications import parse, load, loads, validate
-from ramlfications.models.root import RootNodeAPI08
+from ramlfications.models.raml import RAML08
 from ramlfications.errors import LoadRAMLError
 
 from tests.base import EXAMPLES
@@ -31,7 +31,7 @@ def test_parse(raml):
     config = os.path.join(EXAMPLES + "test-config.ini")
     result = parse(raml, config)
     assert result
-    assert isinstance(result, RootNodeAPI08)
+    assert isinstance(result, RAML08)
 
 
 def test_parse_nonexistant_file():
